requirements.txt
telethon
gspread
apscheduler
python-dotenv
sqlalchemy
google-auth
